Put chat id with Integer as Id and print values

+++ oci-react-samples/MtdrSpring/backend/src/main/java/com/springboot/MyTodoList/repository/TelegramUserRepository.java
@@ -29,10 +29,10 @@ public interface TelegramUserRepository extends JpaRepository<TelegramUser, Long
 
     // get chat id by telegram user id
     @Query(value = "SELECT chatId FROM TODOUSER.TELEGRAMUSER WHERE telegramUserId = ?1", nativeQuery = true)
-    Long fndChatIdByTelegramUserId(Integer id);
+    Long fndChatIdByTelegramUserId(Long id);
 
     // post chat id by telegram user id
     @Query(value = "UPDATE TODOUSER.TELEGRAMUSER SET chatId = ?2 WHERE telegramUserId = ?1", nativeQuery = true)
-    void updateChatIdByTelegramUserId(Integer id, Long chatId);
+    void updateChatIdByTelegramUserId(Long id, Long chatId);
 
 }
